can add new card to frontend, but no validation yet
