+++ src/components/modals/AddCardModal.js
@@ -1,33 +1,53 @@
 import React, { Component } from 'react'
-import bindActionCreators from 'redux'
+import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
+import { addProjectCard } from '../../actions'
 
 class AddCardModal extends Component {
   state = {}
 
   closeModal = () => {
     const modalAddCard = document.querySelector('.modal-add-card')
-
     modalAddCard.classList.add('hide')
+  }
+
+  addCard = (e) => {
+    e.preventDefault()
+
+    const newCardTitle = document.querySelector('#card-title').value
+    const newCardQuantity = document.querySelector('#card-quantity').value
+
+    console.log(newCardTitle, newCardQuantity, "new stuff");
+
+    const newCard = {
+      project_id: 1,
+      quantity: newCardQuantity,
+      properties: [
+        { field_id: 'title', content: newCardTitle }
+      ]
+    }
+
+    this.props.addProjectCard(newCard)
+    this.closeModal()
   }
 
   render () {
     return (
       <div className="modal-wrapper modal-add-card hide">
 
-        <div className="dark-overlay" onClick={this.closeModal}>
+        <div className="dark-overlay" onClick={ this.closeModal }>
         </div>
 
         <div className="modal">
           <div className="modal-close">
             <span>Add New Card</span>
-            <i className="material-icons close-icon" onClick={this.closeModal}>close</i>
+            <i className="material-icons close-icon" onClick={ this.closeModal }>close</i>
           </div>
 
           <div className="mtb-2">
             <div>
               <h3>Title</h3>
-              <input type="text" name="cardTitle"></input>
+              <input type="text" name="cardTitle" id="card-title"></input>
             </div>
             <div className="checkbox-option mtb-05">
               <i className="material-icons small-icon mr-05">check_box</i>
@@ -37,10 +57,10 @@ class AddCardModal extends Component {
 
           <div className="mtb-2">
             <h3>Quantity</h3>
-            <input type="text" name="cardQuantity"></input>
+            <input type="text" name="cardQuantity" id="card-quantity"></input>
           </div>
 
-          <button className="green">Add New Card</button>
+          <button onClick={ this.addCard } className="green">Add New Card</button>
 
         </div>
 
@@ -52,7 +72,9 @@ class AddCardModal extends Component {
 
 const mapStateToProps = state => ({})
 
-const mapDispatchToProps = dispatch => ({})
+const mapDispatchToProps = dispatch => {
+  return bindActionCreators({ addProjectCard }, dispatch)
+}
 
 export default connect(
   mapStateToProps,
