+++ src/components/sidebars/CardActions.js
@@ -1,27 +1,11 @@
 import React from 'react'
-import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import { addProjectCard } from '../../actions'
 
-const CardActions = ( { addProjectCard } ) => {
+const CardActions = ( ) => {
 
-  const addCard = () => {
+  const openAddCardModal = () => {
     const modalAddCard = document.querySelector('.modal-add-card')
-
     modalAddCard.classList.remove('hide')
-
-    const newCard = {
-      _id: 3,
-      name: 'A NEW card',
-      project_id: 1,
-      quantity: 5,
-      properties: [
-        { field_id: 'title', content: 'This is NEW!' },
-        { name: 'Cost', field_id: 'B', content: '10 :coin:' }
-      ]
-    }
-
-    addProjectCard(newCard)
   }
 
   return (
@@ -34,7 +18,7 @@ const CardActions = ( { addProjectCard } ) => {
         <i className="material-icons">check_box_outline_blank</i>
         <span>Unselect All</span>
       </div>
-      <div onClick={addCard} className="mtb-05 local-action" id="add-card">
+      <div onClick={ openAddCardModal } className="mtb-05 local-action" id="add-card">
         <i className="material-icons">add</i>
         <span>Add Card</span>
       </div>
@@ -52,9 +36,7 @@ const CardActions = ( { addProjectCard } ) => {
 
 const mapStateToProps = state => ({})
 
-const mapDispatchToProps = dispatch => {
-  return bindActionCreators({ addProjectCard }, dispatch)
-}
+const mapDispatchToProps = dispatch => ({})
 
 export default connect(
   mapStateToProps,
